+++ src/config.cc
@@ -27,6 +27,11 @@ char *start_with_daily_forecast_values[] = {
     const_cast<char *>(GetLangText("Hourly forecast")),
     const_cast<char *>(GetLangText("Daily forecast")), nullptr};
 
+char *data_update_strategy_values[] = {
+    const_cast<char *>(GetLangText("At startup")),
+    const_cast<char *>(GetLangText("When obsolete")),
+    const_cast<char *>(GetLangText("Manual")), nullptr};
+
 char *custom_api_key_value[] = {const_cast<char *>(""), nullptr};
 
 char *no_yes_values[] = {const_cast<char *>("@No"), const_cast<char *>("@Yes"),
@@ -52,6 +57,10 @@ static iconfigedit main_config_template[] = {
      const_cast<char *>(GetLangText("Start screen")), nullptr,
      const_cast<char *>("start_with_daily_forecast"), const_cast<char *>("0"),
      start_with_daily_forecast_values, nullptr},
+    {CFG_INDEX, &icon_refresh,
+     const_cast<char *>(GetLangText("Data update strategy")), nullptr,
+     const_cast<char *>("data_update_strategy"), const_cast<char *>("0"),
+     data_update_strategy_values, nullptr},
     {CFG_TEXT, &icon_key, const_cast<char *>(GetLangText("Custom API key")),
      nullptr, const_cast<char *>("api_key"), const_cast<char *>(""),
      custom_api_key_value, nullptr},